# wallpaper-server/API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: 壁纸系统API
  description: |
    壁纸系统后端接口文档
    
    ## 认证说明
    
    ### JWT Token认证
    系统使用JWT Token进行身份认证。登录成功后会返回Token，后续请求需要在请求头中携带Token。
    
    **请求头设置：**
    ```
    Authorization: Bearer <token>
    ```
    
    **示例：**
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjEsInVzZXJuYW1lIjoiYWRtaW4iLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MDAwMDAwMDAsImV4cCI6MTcwMDA4NjQwMH0.xxxxx
    ```
    
    **Token有效期：**
    - 默认有效期：24小时（86400000ms）
    - 本地开发环境：7天（604800000ms）
    
    **Token过期处理：**
    - Token过期后，接口会返回 `401` 错误码
    - 错误信息：`登录已过期，请重新登录`
    - 需要重新调用登录接口获取新Token
    
    ### 兼容性说明
    为了兼容旧版本，系统同时支持 `X-User-Id` 请求头（仅开发环境使用）。
    
    ## 错误码说明
    
    ### 成功 (200)
    - 200: 操作成功
    
    ### 用户相关错误 (1000-1099)
    - 1001: 用户不存在
    - 1002: 用户已存在
    - 1003: 用户名或密码错误
    - 1004: 微信未绑定
    - 1005: 用户无权限
    - 1006: 用户暂无创作权限
    
    ### 壁纸相关错误 (2000-2099)
    - 2001: 壁纸不存在
    - 2002: 壁纸上传失败
    - 2003: 壁纸删除失败
    - 2004: 壁纸状态无效
    
    ### 创作者相关错误 (3000-3099)
    - 3001: 创作者不存在
    
    ### 分类相关错误 (4000-4099)
    - 4001: 分类不存在
    - 4002: 分类已存在
    
    ### 专辑相关错误 (5000-5099)
    - 5001: 专辑不存在
    - 5002: 专辑创建失败
    - 5003: 专辑删除失败
    
    ### 作品集相关错误 (8000-8099)
    - 8001: 作品集不存在
    - 8002: 作品集创建失败
    - 8003: 无权操作此作品集
    
    ### 用户行为相关错误 (6000-6099)
    - 6001: 已操作过
    - 6002: 操作记录不存在
    
    ### 举报相关错误 (7000-7099)
    - 7001: 已举报过
    - 7002: 举报失败
    
    ### 参数相关错误 (8000-8099)
    - 8001: 参数错误
    - 8002: 缺少必要参数
...
servers:
...
paths:
...
  /api/user/{id}:
    get:
      tags:
        - 用户管理
      summary: 获取用户信息和统计
      description: 根据用户ID获取用户信息和统计数据（包含作品数、点赞数、收藏数等）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID【必填】
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/UserStats'
...
  /api/category/{id}:
    get:
      tags:
        - 分类管理
      summary: 获取分类基本信息
      description: 根据分类ID获取分类基本信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 分类ID【必填】
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Category'
    put:
      tags:
        - 分类管理
      summary: 更新分类
      description: 根据分类ID更新分类信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 分类ID【必填】
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 分类名称
                  minLength: 2
                  maxLength: 50
                alias:
                  type: string
                  description: 分类别名
                  maxLength: 50
                iconPath:
                  type: string
                  description: 图标路径
                  maxLength: 100
                coverUrl:
                  type: string
                  description: 封面URL
                  maxLength: 500
                sortOrder:
                  type: integer
                  description: 排序值
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Category'
    delete:
      tags:
        - 分类管理
      summary: 删除分类
      description: 根据分类ID删除分类
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 分类ID【必填】
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
...
  /api/works:
    post:
      tags:
        - 作品集
      summary: 创建作品集
      description: 创建新的作品集并上传壁纸，需要登录。管理员上传的作品集会直接发布（无需审核），普通用户上传的作品集需要管理员审核
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - type
              properties:
                title:
                  type: string
                  description: 作品集标题
                  example: 动漫壁纸合集
                type:
                  type: integer
                  description: 作品类型（1手机壁纸、2平板壁纸、3头像）
                  example: 1
                categoryId:
                  type: integer
                  description: 分类ID
                  example: 1
                description:
                  type: string
                  description: 作品集描述
                  example: 精选动漫壁纸合集
                wallpapers:
                  type: array
                  description: 壁纸信息列表（至少需要一张）
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                        description: 壁纸原始图片URL
                        example: https://example.com/wallpaper1.jpg
                      thumbnailUrl:
                        type: string
                        description: 壁纸缩略图URL
                        example: https://example.com/thumb1.jpg
                      imageWidth:
                        type: integer
                        description: 图片宽度（像素）
                        example: 1920
                      imageHeight:
                        type: integer
                        description: 图片高度（像素）
                        example: 1080
                      fileFormat:
                        type: string
                        description: 图片格式（JPG、PNG等）
                        example: JPG
                      fileSize:
                        type: integer
                        description: 文件大小（字节）
                        example: 1024000
                  minItems: 1
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Work'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/works/{id}:
    get:
      tags:
        - 作品集
      summary: 获取作品集详情
      description: 根据ID获取作品集详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Work'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'

    put:
      tags:
        - 作品集
      summary: 更新作品集
      description: 更新作品集信息，需要登录且是作品集作者
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 作品集标题
                  example: 动漫壁纸合集
                type:
                  type: integer
                  description: 作品类型（1手机壁纸、2平板壁纸、3头像）
                  example: 1
                categoryId:
                  type: integer
                  description: 分类ID
                  example: 1
                description:
                  type: string
                  description: 作品集描述
                  example: 精选动漫壁纸合集
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Work'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8003':
          description: 无权修改此作品集
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'

    delete:
      tags:
        - 作品集
      summary: 删除作品集
      description: 删除作品集，需要登录且是作品集作者
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8003':
          description: 无权删除此作品集
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/works/user/{userId}:
    get:
      tags:
        - 作品集
      summary: 获取用户的作品集
      description: 获取指定用户的作品集列表
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: 页码
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: 每页数量
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/Work'
                      total:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
  /api/works/category/{categoryId}:
    get:
      tags:
        - 作品集
      summary: 获取分类下的作品集
      description: 获取指定分类下的作品集列表
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
          description: 分类ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: 页码
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: 每页数量
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      list:
                        type: array
                        items:
                          $ref: '#/components/schemas/Work'
                      total:
                        type: integer
                        example: 20
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
  /api/works/{workId}/wallpapers:
    get:
      tags:
        - 作品集
      summary: 获取作品集的壁纸列表
      description: 获取指定作品集下的所有壁纸
      parameters:
        - name: workId
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
        - name: userId
          in: query
          required: false
          schema:
            type: integer
          description: 当前用户ID（用于判断是否喜欢/收藏）
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wallpaper'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/works/{workId}/like:
    post:
      tags:
        - 作品集
      summary: 点赞/取消点赞作品集
      description: 对作品集进行点赞或取消点赞，需要登录
      security:
        - bearerAuth: []
      parameters:
        - name: workId
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/works/{workId}/collect:
    post:
      tags:
        - 作品集
      summary: 收藏/取消收藏作品集
      description: 对作品集进行收藏或取消收藏，需要登录
      security:
        - bearerAuth: []
      parameters:
        - name: workId
          in: path
          required: true
          schema:
            type: integer
          description: 作品集ID
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '8001':
          description: 作品集不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证，格式为 `Bearer <token>`
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
          description: 状态码（200成功，其他为错误码）
        message:
          type: string
          example: success
          description: 返回消息
        data:
          type: object
          nullable: true
          description: 返回数据
    UserStats:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        nickname:
          type: string
          example: 小明
        avatar:
          type: string
          example: /static/images/avatar1.jpg
        role:
          type: string
          example: user
        isCreator:
          type: boolean
          example: true
        followerCount:
          type: integer
          example: 100
        workCount:
          type: integer
          example: 10
        totalLikes:
          type: integer
          example: 500
        totalCollects:
          type: integer
          example: 200
        bio:
          type: string
          example: 个人简介
...
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 动漫次元
        alias:
          type: string
          example: anime
        iconPath:
          type: string
          example: /static/icons/anime.png
        coverUrl:
          type: string
          example: /static/images/category_cover.jpg
        likeCount:
          type: integer
          example: 100
        sortOrder:
          type: integer
          example: 1
        createTime:
          type: string
          format: date-time
          example: 2026-03-16T10:00:00
        updateTime:
          type: string
          format: date-time
          example: 2026-03-16T10:00:00
        isDel:
          type: boolean
          example: false
...
    Work:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: 动漫壁纸合集
        coverUrl:
          type: string
          description: 封面地址（第一张壁纸）
          example: /static/images/wallpaper1.jpg
        type:
          type: integer
          description: 作品类型（1手机壁纸、2平板壁纸、3头像）
          example: 1
        categoryId:
          type: integer
          description: 分类ID
          example: 1
        description:
          type: string
          description: 作品集描述
          example: 精选动漫壁纸合集
        userId:
          type: integer
          description: 用户ID
          example: 1
        wallpaperCount:
          type: integer
          description: 作品数量
          example: 10
        likeCount:
          type: integer
          description: 点赞次数
          example: 100
        collectCount:
          type: integer
          description: 收藏次数
          example: 50
        viewCount:
          type: integer
          description: 浏览次数
          example: 500
        downloadCount:
          type: integer
          description: 下载次数
          example: 200
        status:
          type: integer
          description: 审核状态（0待审核、1已通过、2拒绝）
          example: 1
        createTime:
          type: string
          format: date-time
          example: 2026-03-16T10:00:00
        updateTime:
          type: string
          format: date-time
          example: 2026-03-16T10:00:00
        isDel:
          type: boolean
          description: 是否删除
          example: false
...